web_app/requirements.txt
Flask==2.3.3
psycopg[binary]==3.2.11
pandas==2.3.3
python-dotenv==1.0.0
numpy==2.3.4